refactor core types a bit
