+++ web/src/beta/lib/core/engines/Cesium/Feature/context.ts
@@ -2,12 +2,13 @@ import { createContext, useContext as useReactContext } from "react";
 
 import { LayerEditEvent } from "@reearth/beta/lib/core/Map";
 
-import type { Camera, FlyToDestination, CameraOptions, LayerSelectionReason } from "../..";
+import type { Camera, LayerSelectionReason } from "../..";
+import type { FlyTo } from "../../../types";
 
 export type Context = {
   selectionReason?: LayerSelectionReason;
   getCamera?: () => Camera | undefined;
-  flyTo?: (target: string | FlyToDestination, options?: CameraOptions) => void;
+  flyTo?: FlyTo;
   onLayerEdit?: (e: LayerEditEvent) => void;
   requestRender?: () => void;
 };
